--- src/main/java/poomasi/domain/auth/blacklist/repository/BlacklistRepository.java
@@ -1,4 +1,0 @@
-package poomasi.domain.auth.blacklist.repository;
-
-public interface BlacklistRepository {
-}
