--- src/main/java/poomasi/domain/auth/blacklist/service/TokenBlacklistService.java
@@ -1,4 +1,0 @@
-package poomasi.domain.auth.blacklist.service;
-
-public class TokenBlacklistService {
-}
