--- src/main/java/poomasi/domain/auth/reissue/dto/request/ReissueTokenRequest.java
@@ -1,8 +1,0 @@
-package poomasi.domain.auth.reissue.dto.request;
-
-public record ReissueTokenRequest(String refreshToken) {
-
-
-
-
-}
