--- src/main/java/poomasi/domain/auth/reissue/dto/response/RefreshTokenResponse.java
@@ -1,4 +1,0 @@
-package poomasi.domain.auth.reissue.dto.response;
-
-public record RefreshTokenResponse() {
-}
