+++ src/main/java/poomasi/domain/auth/signup/controller/SignupController.java
@@ -3,6 +3,7 @@ package poomasi.domain.auth.signup.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import poomasi.domain.auth.signup.dto.response.SignUpResponse;
 import poomasi.domain.auth.signup.service.SignupService;
 import poomasi.domain.auth.signup.dto.request.SignupRequest;
 
@@ -14,7 +15,7 @@ public class SignupController {
     private final SignupService signupService;
 
     @PostMapping("/sign-up")
-    public ResponseEntity<?> signUp(@RequestBody SignupRequest signupRequest) {
+    public ResponseEntity<SignUpResponse> signUp(@RequestBody SignupRequest signupRequest) {
         return ResponseEntity.ok(signupService
                 .signUp(signupRequest));
     }
