+++ src/main/java/poomasi/domain/auth/token/blacklist/repository/BlacklistRepository.java
@@ -1,0 +1,4 @@
+package poomasi.domain.auth.token.blacklist.repository;
+
+public interface BlacklistRepository {
+}
