+++ src/main/java/poomasi/domain/auth/token/blacklist/service/TokenBlacklistService.java
@@ -1,0 +1,4 @@
+package poomasi.domain.auth.token.blacklist.service;
+
+public class TokenBlacklistService {
+}
