+++ src/main/java/poomasi/domain/auth/token/dto/response/TokenResponse.java
@@ -1,0 +1,4 @@
+package poomasi.domain.auth.token.dto.response;
+
+public record TokenResponse(String accessToken, String refreshToken) {
+}
