+++ src/main/java/poomasi/domain/auth/token/reissue/controller/ReissueTokenController.java
@@ -1,0 +1,23 @@
+package poomasi.domain.auth.token.reissue.controller;
+
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import poomasi.domain.auth.token.dto.response.TokenResponse;
+import poomasi.domain.auth.token.reissue.dto.ReissueRequest;
+import poomasi.domain.auth.token.reissue.service.ReissueTokenService;
+
+@RestController
+public class ReissueTokenController {
+
+    @Autowired
+    private ReissueTokenService reissueTokenService;
+
+    @GetMapping("/api/reissue")
+    public ResponseEntity<TokenResponse> reissue(@RequestBody ReissueRequest reissueRequest){
+        return ResponseEntity.ok(reissueTokenService.reissueToken(reissueRequest));
+    }
+}
