+++ src/main/java/poomasi/domain/auth/token/reissue/dto/ReissueRequest.java
@@ -1,0 +1,4 @@
+package poomasi.domain.auth.token.reissue.dto;
+
+public record ReissueRequest(String refreshToken) {
+}
