+++ src/main/java/poomasi/global/redis/error/RedisConnectionException.java
@@ -4,8 +4,4 @@ public class RedisConnectionException extends RuntimeException {
     public RedisConnectionException(String message) {
         super(message);
     }
-
-    public RedisConnectionException(String message, Throwable cause) {
-        super(message, cause);
-    }
 }
