+++ src/main/java/poomasi/global/redis/error/RedisOperationException.java
@@ -4,8 +4,4 @@ public class RedisOperationException extends RuntimeException {
     public RedisOperationException(String message) {
         super(message);
     }
-
-    public RedisOperationException(String message, Throwable cause) {
-        super(message, cause);
-    }
 }
